+++ samples/bot-calling-meeting/csharp/.gitignore
@@ -458,4 +458,4 @@ $RECYCLE.BIN/
 ##
 ## Microsoft Teams development
 ##
-Manifest/*.zip
+**/Manifest/*.zip
