+++ samples/bot-calling-meeting/csharp/Source/CallingBotSample/AdapterWithErrorHandler.cs
