+++ samples/bot-calling-meeting/csharp/Source/CallingBotSample/Authentication/AuthenticationWrapper.cs
