+++ samples/bot-calling-meeting/csharp/Source/CallingBotSample/Bots/CallingBot.cs
