+++ samples/bot-calling-meeting/csharp/Source/CallingBotSample/CallingBotSample.csproj
