+++ samples/bot-calling-meeting/csharp/Source/CallingBotSample/Common/Constants.cs
