+++ samples/bot-calling-meeting/csharp/Source/CallingBotSample/Configuration/Users.cs
