+++ samples/bot-calling-meeting/csharp/Source/CallingBotSample/Extensions/GraphConfiguration.cs
