+++ samples/bot-calling-meeting/csharp/Source/CallingBotSample/Extensions/HttpExtensions.cs
