+++ samples/bot-calling-meeting/csharp/Source/CallingBotSample/Helpers/CardHelper.cs
