+++ samples/bot-calling-meeting/csharp/Source/CallingBotSample/Helpers/GraphHelper.cs
