+++ samples/bot-calling-meeting/csharp/Source/CallingBotSample/Helpers/JoinInfo.cs
