+++ samples/bot-calling-meeting/csharp/Source/CallingBotSample/Interfaces/ICard.cs
