+++ samples/bot-calling-meeting/csharp/Source/CallingBotSample/Interfaces/IGraph.cs
