+++ samples/bot-calling-meeting/csharp/Source/CallingBotSample/Manifest/color.png
