+++ samples/bot-calling-meeting/csharp/Source/CallingBotSample/Manifest/manifest.json
