+++ samples/bot-calling-meeting/csharp/Source/CallingBotSample/Manifest/outline.png
