+++ samples/bot-calling-meeting/csharp/Source/CallingBotSample/Model/BotOptions.cs
