+++ samples/bot-calling-meeting/csharp/Source/CallingBotSample/Model/ParticipantDetails.cs
