+++ samples/bot-calling-meeting/csharp/Source/CallingBotSample/Properties/launchSettings.json
