+++ samples/bot-calling-meeting/csharp/Source/CallingBotSample/Resources/WelcomeCard.json
