+++ samples/bot-calling-meeting/csharp/Source/CallingBotSample/Utility/CommonUtils.cs
