+++ samples/bot-calling-meeting/csharp/Source/CallingBotSample/appsettings.json
