+++ samples/bot-calling-meeting/csharp/Source/CallingBotSample/wwwroot/audio/speech.wav
