+++ samples/bot-calling-meeting/csharp/Source/CallingBotSample/wwwroot/default.htm
